+++ components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/composable/ComposableInAppNotification.kt
@@ -1,12 +1,12 @@
 package com.flipperdevices.bottombar.impl.composable
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import com.flipperdevices.bottombar.impl.viewmodel.InAppNotificationState
 import com.flipperdevices.bottombar.impl.viewmodel.InAppNotificationViewModel
+import com.flipperdevices.core.ktx.android.OnLifecycleEvent
 import com.flipperdevices.inappnotification.api.InAppNotificationRenderer
 import tangle.viewmodel.compose.tangleViewModel
 
@@ -18,12 +18,7 @@ fun ComposableInAppNotification(
 ) {
     val notificationState by notificationViewModel.state().collectAsState()
 
-    DisposableEffect(key1 = Unit) {
-        notificationViewModel.onCreate()
-        onDispose {
-            notificationViewModel.onDestroy()
-        }
-    }
+    OnLifecycleEvent(onEvent = notificationViewModel::onLifecycleEvent)
 
     val localNotificationState = notificationState
     if (localNotificationState !is InAppNotificationState.ShownNotification) {