+++ components/faphub/fapscreen/impl/src/main/java/com/flipperdevices/faphub/fapscreen/impl/api/FapScreenApiImpl.kt
@@ -11,7 +11,6 @@ import com.flipperdevices.core.ui.navigation.ComposableFeatureEntry
 import com.flipperdevices.faphub.fapscreen.api.FapScreenApi
 import com.flipperdevices.faphub.fapscreen.impl.composable.ComposableFapScreen
 import com.flipperdevices.faphub.installation.api.FapInstallationUIApi
-import com.flipperdevices.faphub.search.api.FapHubSearchEntryApi
 import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
@@ -21,7 +20,6 @@ internal const val FAP_ID_KEY = "fap_id"
 @ContributesBinding(AppGraph::class, FapScreenApi::class)
 @ContributesMultibinding(AppGraph::class, ComposableFeatureEntry::class)
 class FapScreenApiImpl @Inject constructor(
-    private val searchEntryApi: FapHubSearchEntryApi,
     private val installationUIApi: FapInstallationUIApi
 ) : FapScreenApi {
     private val fapArguments = listOf(
@@ -42,9 +40,12 @@ class FapScreenApiImpl @Inject constructor(
         ) {
             ComposableFapScreen(
                 onBack = navController::popBackStack,
-                onSearch = { navController.navigate(searchEntryApi.start()) },
                 installationButton = { fapItem, modifier, fontSize ->
-                    installationUIApi.ComposableInstallButton(fapItem, modifier, fontSize)
+                    installationUIApi.ComposableButton(
+                        fapItem = fapItem,
+                        modifier = modifier,
+                        textSize = fontSize
+                    )
                 }
             )
         }