+++ components/selfupdater/source/googleplay/src/main/kotlin/com/flipperdevices/selfupdater/googleplay/api/SelfUpdaterGooglePlayApi.kt
@@ -7,6 +7,7 @@ import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
 import com.flipperdevices.inappnotification.api.InAppNotificationStorage
 import com.flipperdevices.inappnotification.api.model.InAppNotification
+import com.flipperdevices.selfupdater.api.BuildConfig
 import com.flipperdevices.selfupdater.api.SelfUpdaterApi
 import com.google.android.play.core.appupdate.AppUpdateInfo
 import com.google.android.play.core.appupdate.AppUpdateManagerFactory
@@ -60,7 +61,7 @@ class SelfUpdaterGooglePlayApi @Inject constructor(
         }
     }
 
-    override fun getInstallSourceName() = "Google Play"
+    override fun getInstallSourceName() = "Google Play/" + BuildConfig.BUILD_TYPE
 
     private fun isUpdateAvailable(appUpdateInfo: AppUpdateInfo): Boolean {
         return appUpdateInfo.updateAvailability() == UpdateAvailability.UPDATE_AVAILABLE &&