+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/category/DebugCategory.kt
@@ -1,6 +1,5 @@
 package com.flipperdevices.settings.impl.composable.category
 
-import androidx.compose.foundation.layout.Column
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
@@ -24,16 +23,14 @@ fun DebugCategory(
     debugViewModel: DebugViewModel = tangleViewModel()
 ) {
     CardCategory(modifier = modifier) {
-        Column {
-            CategoryElement(
-                titleId = R.string.debug_options,
-                descriptionId = R.string.debug_options_desc,
-                state = settings.enabledDebugSettings,
-                onSwitchState = settingsViewModel::onSwitchDebug
-            )
-            if (settings.enabledDebugSettings) {
-                DebugCategoryItems(navController, debugViewModel, settings)
-            }
+        CategoryElement(
+            titleId = R.string.debug_options,
+            descriptionId = R.string.debug_options_desc,
+            state = settings.enabledDebugSettings,
+            onSwitchState = settingsViewModel::onSwitchDebug
+        )
+        if (settings.enabledDebugSettings) {
+            DebugCategoryItems(navController, debugViewModel, settings)
         }
     }
 }
@@ -91,5 +88,11 @@ private fun DebugCategoryItems(
     ClickableElement(
         titleId = R.string.debug_open_mfkey32,
         onClick = { debugViewModel.openMfKey32(navController) }
+    )
+    GrayDivider()
+    SwitchableElement(
+        titleId = R.string.experimental_application_catalog,
+        state = settings.applicationCatalog,
+        onSwitchState = debugViewModel::onSwitchApplicationCatalog
     )
 }