+++ CHANGELOG.md
@@ -13,11 +13,15 @@
 - [Feature] New report bug screen
 - [Feature] Self update(Google play)
 - [Feature] Add SAVE_DUMP, MFKEY32, OPEN_NFC_DUMP_EDITOR
+- [Feature] Add pull to refresh to full device info
+- [Feature] Add report changes to synchronization end event
+- [Feature] Add GitHub and Forum link to options
 - [GitHub] Update github repo picture
 - [CI] Migrate to Detekt
 - [CI] Android lint in CI
 - [CI] Enable back github queue
 - [CI] Migrate to matrix CI with source install
+- [CI] Use release build for release
 - [FIX] Clickable modifier ktx
 - [FIX] Refactor `share:receive` module
 - [FIX] Bump length for server-side share link
@@ -32,9 +36,17 @@
 - [FIX] Fix change color scheme bug
 - [FIX] Fix status bar color
 - [FIX] Fix open key screen in global scope
+- [FIX] Fix fap hub design
 - [FIX] Wait for disconnect while update
 - [FIX] Fix mfkey32
 - [FIX] Use CurrentActivityHolder for get activity in updater(replace Context cast)
+- [FIX] Back on mfkey32 screen
+- [FIX] Cash with ulong mapping to string in mfkey32
+- [FIX] Move apps from experimental settings to debug
+- [FIX] Fix customize button when set widget state
+- [FIX] Fix incorrect diff combiner, add KeyDiffCombinerTest
+- [FIX] Use lifecycle event for inapp notification
+- [FIX] Fix Report bug scroll
 - [REFACTOR] Migrate bottom bar to compose navigation
 - [REFACTOR] Bump Android Gradle Plugin
 - [REFACTOR] Fix detekt compose issues